feat(issue-165): consolidate python micro-modules
