+++ python_tools/__init__.py
@@ -1,2 +1,1 @@
 #!/usr/bin/env python3
-
