+++ python_tools/ci/fmea_status.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 import json
 import shutil
+from collections.abc import Mapping, Sequence
 from datetime import UTC, datetime
 from pathlib import Path
 
-from python_tools.ci.fmea_status_assets import render_fmea_status_readme
 from python_tools.policies.fmea_action_register import FmeaActionRegister
 
 
@@ -43,3 +43,23 @@ class FmeaStatusSnapshot:
         (dist_dir / "README.md").write_text(render_fmea_status_readme(snapshot), encoding="utf-8")
         archive_base = dist_dir / "CUDA-GRAVITY-SIMULATION-fmea-status"
         return Path(shutil.make_archive(str(archive_base), "zip", root_dir=dist_dir))
+
+
+def render_fmea_status_readme(snapshot: Mapping[str, object]) -> str:
+    summary = snapshot.get("summary")
+    if not isinstance(summary, dict):
+        summary = {}
+    high_medium = snapshot.get("high_medium")
+    if not isinstance(high_medium, Sequence):
+        high_medium = ()
+    lines = [
+        "# FMEA Risk Status Snapshot",
+        "",
+        f"- Open: `{summary.get('open', 0)}`",
+        f"- In Progress: `{summary.get('in_progress', 0)}`",
+        f"- Closed: `{summary.get('closed', 0)}`",
+        f"- High/Medium residual risks: `{len(high_medium)}`",
+        "",
+        "See `fmea_status_snapshot.json` for machine-readable details.",
+    ]
+    return "\n".join(lines) + "\n"
