+++ python_tools/ci/release_support.py
@@ -2,6 +2,36 @@
 from __future__ import annotations
 
 import os
+from collections.abc import Mapping
+from pathlib import Path
+
+from python_tools.policies.deviation_register import DeviationRegister
+
+DEFAULT_EVIDENCE_REFS = (
+    "EVD_CI_RELEASE_LANE",
+    "EVD_QLT_DEVIATION_REGISTER",
+    "EVD_QLT_EVIDENCE_PACK_FORMAT",
+    "EVD_QLT_MANIFEST",
+    "EVD_QLT_PROD_BASELINE",
+    "EVD_QLT_README",
+    "EVD_QLT_STANDARDS_PROFILE",
+    "EVD_SCRIPT_RELEASE_PACKAGE_BUNDLE",
+    "EVD_SCRIPT_RELEASE_PACKAGE_EVIDENCE",
+)
+
+
+def resolve_release_tag(explicit: str | None) -> str:
+    if explicit is not None and explicit.strip():
+        return explicit.strip()
+    ref_name = os.environ.get("GITHUB_REF_NAME", "").strip()
+    run_number = os.environ.get("GITHUB_RUN_NUMBER", "").strip()
+    if ref_name and run_number:
+        return f"{ref_name}-{run_number}"
+    if ref_name:
+        return ref_name
+    if run_number:
+        return f"run-{run_number}"
+    return "manual"
 
 
 def build_release_lane_activities(profile: str) -> list[dict[str, str]]:
@@ -51,3 +81,28 @@ def default_ci_context() -> dict[str, str]:
     if server_url and context["repository"] and context["run_id"]:
         context["run_url"] = f"{server_url}/{context['repository']}/actions/runs/{context['run_id']}"
     return {key: value for key, value in context.items() if value}
+
+
+def load_open_exceptions(root: Path) -> list[dict[str, object]]:
+    rows, errors = DeviationRegister().load_open_with_errors(root)
+    if errors:
+        raise RuntimeError(errors[0])
+    return [row for row in rows if isinstance(row, dict)]
+
+
+def render_pack_readme(pack: Mapping[str, object]) -> str:
+    requirement_ids = pack.get("requirement_ids")
+    evidence_refs = pack.get("evidence_refs")
+    open_exceptions = pack.get("open_exceptions")
+    requirement_count = len(requirement_ids) if isinstance(requirement_ids, list) else 0
+    evidence_count = len(evidence_refs) if isinstance(evidence_refs, list) else 0
+    exception_count = len(open_exceptions) if isinstance(open_exceptions, list) else 0
+    return (
+        "# Release Evidence Pack\n\n"
+        f"- Tag: `{pack.get('tag', '')}`\n"
+        f"- Profile: `{pack.get('profile', '')}`\n"
+        f"- Requirements covered: `{requirement_count}`\n"
+        f"- Evidence refs bundled: `{evidence_count}`\n"
+        f"- Open exceptions recorded: `{exception_count}`\n\n"
+        "See `release_evidence_pack.json` for the machine-readable record.\n"
+    )
