--- python_tools/core/contracts.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-from __future__ import annotations
-
-from typing import Protocol
-
-from .models import CheckContext, CheckResult
-
-
-class CheckContract(Protocol):
-    def run(self, context: CheckContext) -> CheckResult:
-        ...
-
